Create Java Static Initializer Block
There are two lines of input. The first line contains B: the breadth of the parallelogram. The next line contains H: the height of the parallelogram. If both values are greater than zero, then the main method must output the area of the parallelogram. Otherwise, print "java.lang.Exception: Breadth and height must be positive" without quotes.