Create Java Anagrams
Two strings, a and b, are called anagrams if they contain all the same characters in the same frequencies. For this challenge, the test is not case-sensitive. For example, the anagrams of CAT are CAT, ACT, tac, TCA, aTC, and CtA. If a and b are case-insensitive anagrams, return true. Otherwise, return false.