Create Pattern Syntax Checker
Using Regex, we can easily match or search for patterns in a text. Before searching for a pattern, we have to specify one using some well-defined syntax.
In this problem, you are given a pattern. You have to check whether the syntax of the given pattern is valid..Print Valid if the syntax of the given pattern is correct. Otherwise, print Invalid.