Create Java Regex
You need to write a regular expression and assign it to the pattern such that it can be used to validate an IP address. In this problem, you will be provided strings containing any combination of ASCII characters. You have to write a regular expression to find the valid IPs.